Move general createRelationshipFilter from Neo4jUtil to QueryUtils
